[FLINK-21927][table-planner-blink] Resolve ExpressionReducer compile fail when running long term (#15354)
This commit replaces CodeGenUtils#nameCounter from AtomicInteger to AtomicLong to resolve
Integer overflow issue

+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/codegen/CodeGenUtils.scala
@@ -20,7 +20,7 @@ package org.apache.flink.table.planner.codegen
 
 import java.lang.reflect.Method
 import java.lang.{Boolean => JBoolean, Byte => JByte, Double => JDouble, Float => JFloat, Integer => JInt, Long => JLong, Object => JObject, Short => JShort}
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.functions.RuntimeContext
@@ -116,7 +116,7 @@ object CodeGenUtils {
 
   // ----------------------------------------------------------------------------------------
 
-  private val nameCounter = new AtomicInteger
+  private val nameCounter = new AtomicLong
 
   def newName(name: String): String = {
     s"$name$$${nameCounter.getAndIncrement}"
